+++ savannah-canopy-rest/src/main/java/com/plants_store/savannah_canopy/controller/PlantController.java
@@ -27,7 +27,7 @@ public class PlantController {
     private static final Logger logger = LoggerFactory.getLogger(PlantController.class);
 
     @Autowired
-    private PlantService plantService;
+    private PlantService plantService; // injecting the interface
 
     @GetMapping("/plant/{id}")
     @Operation(summary = "Get a plant by ID", description = "Retrieves a plant based on its unique identifier")
@@ -41,23 +41,6 @@ public class PlantController {
         logger.info("Plant retrieved successfully");
         return ResponseEntity.ok(plant);
     }
-
-//    public ResponseEntity<Plant> getPlant(@PathVariable Long id) {
-//        MDC.put("plantId", id.toString()); // Log the plant ID
-//        try {
-//            Plant plant = plantService.getPlantByID(id);
-//            if (plant == null) {
-//                throw new IllegalArgumentException("Plant cannnot be found");
-//            }
-//            logger.info("Plant retrieved successfully");
-//            return ResponseEntity.ok(plant);
-//        } catch (Exception ex) {
-//            logger.error("An error occurred while processing the request", ex); //pass the exception object as a parameter to the logger to capture the complete stack trace
-//            throw ex;
-//        } finally {
-//            MDC.clear();
-//        }
-//    }
 
     @GetMapping("/plants")
     @Operation(summary = "Get all plants", description = "Retrieves all plants")
