+++ savannah-canopy-rest/src/main/java/com/plants_store/savannah_canopy/service/PlantService.java
@@ -1,37 +1,31 @@
-package com.plants_store.savannah_canopy.service; 
- 
-import com.plants_store.savannah_canopy.exception.ErrorContext; 
-import com.plants_store.savannah_canopy.model.Plant; 
-import com.plants_store.savannah_canopy.repository.PlantRepository; 
-import org.springframework.beans.factory.annotation.Autowired; 
-import org.springframework.stereotype.Service; 
- 
-import java.util.List; 
- 
-/** 
- * Service layer for handling plant-related operations. 
- */ 
-@Service 
-public class PlantService { 
-    @Autowired 
-    private PlantRepository plantRepository; 
- 
-    // Retrieve plant by id 
-    public Plant getPlantByID(long id){ 
-        Plant plant = plantRepository.findById(id).orElse(null); 
-        if (plant == null) { 
-            throw new ErrorContext("Plant cannot be found",id); 
-        } 
-        return  plant; 
-    } 
- 
-    // Retrieve all plant products 
-    public List<Plant> getAllPlants() { 
-        return plantRepository.findAll(); 
-    } 
- 
-    // Save a plant product 
-    public Plant savePlant(Plant plant) { 
-        return plantRepository.save(plant); 
-    } 
+package com.plants_store.savannah_canopy.service;
+
+import com.plants_store.savannah_canopy.model.Plant;
+import java.util.List;
+
+/**
+ * Interface defining operations for managing plants.
+ */
+public interface PlantService {
+
+    /**
+     * Retrieves a plant by its unique identifier.
+     * @param id The ID of the plant to retrieve.
+     * @return The found Plant object.
+     * @throws com.plants_store.savannah_canopy.exception.ErrorContext if the plant is not found.
+     */
+    Plant getPlantByID(long id);
+
+    /**
+     * Retrieves all available plants.
+     * @return A list of all Plant objects.
+     */
+    List<Plant> getAllPlants();
+
+    /**
+     * Saves a new plant or updates an existing one.
+     * @param plant The Plant object to save.
+     * @return The saved Plant object, potentially with updated state (like ID).
+     */
+    Plant savePlant(Plant plant);
 }
